requirements.txt
asgiref==3.4.1
attrs==22.2.0
certifi==2020.12.5
cffi==1.15.0
chardet==4.0.0
cryptography==36.0.1
Django==3.2.9
django-ckeditor==6.2.0
django-jazzmin==2.6.0
django-js-asset==1.2.2
django-mysql==4.5.0
django-rest-framework==0.1.0
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
et-xmlfile==1.1.0
exceptiongroup==1.1.0
Faker==17.5.0
gunicorn==20.1.0
idna==2.10
iniconfig==2.0.0
model-bakery==1.10.1
numpy==1.22.3
openpyxl==3.1.1
packaging==23.0
Pillow==8.4.0
pluggy==1.0.0
psycopg2-binary==2.9.2
pycparser==2.21
PyMySQL==1.0.2
pytest==7.2.1
pytest-django==4.5.2
python-dateutil==2.8.2
pytz==2020.4
qrcode==7.3.1
requests==2.25.1
six==1.16.0
sqlparse==0.4.1
tomli==2.0.1
urllib3==1.26.4
whitenoise==5.2.0

